requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
amqp==5.1.1; python_version >= '3.6'
appnope==0.1.3; sys_platform == 'darwin'
asgiref==3.7.2; python_version >= '3.7'
asttokens==2.4.0
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
backcall==0.2.0
bcrypt==4.0.1
billiard==4.1.0; python_version >= '3.7'
celery==5.3.4
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
decorator==5.1.1; python_version >= '3.5'
django==4.2.6; python_version >= '3.8'
djangorestframework==3.14.0
executing==2.0.0
frozenlist==1.4.0; python_version >= '3.8'
html2text==2020.1.16
idna==3.4; python_version >= '3.5'
ipython==8.16.1
jedi==0.19.1; python_version >= '3.6'
kombu==5.3.2; python_version >= '3.8'
matplotlib-inline==0.1.6; python_version >= '3.5'
multidict==6.0.4; python_version >= '3.7'
openai==0.28.1
parso==0.8.3; python_version >= '3.6'
passlib==1.7.4
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.39; python_full_version >= '3.7.0'
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.16.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
stack-data==0.6.3
tqdm==4.66.1; python_version >= '3.7'
traitlets==5.11.2; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
urllib3==2.0.6; python_version >= '3.7'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.8
yarl==1.9.2; python_version >= '3.7'
